divided general FB screen compornent
